% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.r
\name{arcsinh_transform}
\alias{arcsinh_transform}
\title{apply arcsinh transform}
\usage{
arcsinh_transform(in_data, a = 1, b = 1, c = 0)
}
\arguments{
\item{in_data}{the data to transform}

\item{a}{first parameter (default = 1)}

\item{b}{second parameter (default = 1)}

\item{c}{third parameter (default = 0)}
}
\value{
matrix
}
\description{
arcsinh is similar to log, but handles zeros and negatives
}
\details{
\code{asinh(a + b * x) + c}
}

