% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.r
\name{filter_features}
\alias{filter_features}
\title{filter data}
\usage{
filter_features(in_data, min_value)
}
\arguments{
\item{in_data}{original input data}

\item{min_value}{what minimum value to filter out}
}
\value{
data.frame
}
\description{
filter data
}

