% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.r
\name{plot_diff}
\alias{plot_diff}
\title{plot differences in range}
\usage{
plot_diff(org_data, log_mean = log1p)
}
\arguments{
\item{org_data}{the original data}

\item{log_mean}{should plot with log-transformed mean on x-axis?}
}
\value{
plot
}
\description{
plot differences in range
}

