% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.r
\name{plot_hist}
\alias{plot_hist}
\title{generate histogram}
\usage{
plot_hist(in_data, bins = 100)
}
\arguments{
\item{in_data}{original data}

\item{bins}{how many bins to use}
}
\value{
plot
}
\description{
generate histogram
}

