% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.r
\name{plot_sd}
\alias{plot_sd}
\title{sd and rsd plot}
\usage{
plot_sd(org_data, log_mean = log1p, sd_type = "sd")
}
\arguments{
\item{org_data}{the original data as matrix}

\item{log_mean}{should the mean on x-axis be log-transformed}

\item{sd_type}{plot the SD or RSD?}
}
\value{
plot
}
\description{
sd and rsd plot
}

