% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.r
\name{root_transform}
\alias{root_transform}
\title{root transform the data}
\usage{
root_transform(in_data, root = 2)
}
\arguments{
\item{in_data}{data to transform}

\item{root}{what root to use in the transformation}
}
\value{
matrix
}
\description{
root transform the data
}

