% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.r
\name{summarize_two}
\alias{summarize_two}
\title{get sd or rsd from two datasets}
\usage{
summarize_two(data1, data2, sd_type = "sd", name1 = "data1",
  name2 = "data2")
}
\arguments{
\item{data1}{the first dataset}

\item{data2}{the second dataset}
}
\value{
data.frame
}
\description{
get sd or rsd from two datasets
}

